bot/requirements.txt
python-dotenv==0.19.0
pytube==11.0.1
python-telegram-bot==13.9
requests